Refactor contract binding generation into smaller modules
